Factor out utils from typecheck.rs
