+++ lang/elaborator/src/typechecker/mod.rs
@@ -1,5 +1,7 @@
 pub mod ctx;
+pub mod subst;
 pub mod typecheck;
+pub mod util;
 
 pub use crate::result::TypeError;
 pub use tracer;
